Pull out abstract parent class for DefaultContextAwareScheduledExecutorService
